refactor: update data types and structure for blog and post models
- Changed `user_id` from `int64` to `int` in session handling
- Refactored `Blog` and `Post` models to use more appropriate data types
- Introduced `BlogView` and `PostView` types for improved presentation
- Updated `PostModel` to include new fields and improve database queries
- Refactored methods to accommodate new structure and improve flexibility

+++ cmd/web/middleware.go
@@ -44,14 +44,14 @@ func (app *application) requireAuthenticatedUser(next http.Handler) http.Handler
 func (app *application) authenticate(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		session, _ := app.session.Get(r, "user-session")
-		userID, exists := session.Values["user_id"].(int64)
+		userID, exists := session.Values["user_id"].(int)
 		if !exists {
 			app.infoLog.Println("session not exist")
 			next.ServeHTTP(w, r)
 			return
 		}
 
-		user, err := app.models.Users.Get(int64(userID))
+		user, err := app.models.Users.Get(int(userID))
 		if err == models.ErrRecordNotFound {
 			app.infoLog.Println("no record find [from auth middleware]")
 			delete(session.Values, "user_id")
